feature: move music table component
